+++ lib/features/game_history/presentation/game_history_bloc/game_history_event.dart
@@ -1,0 +1,9 @@
+import 'package:keno_plus/features/game_history/data/models/game_history_model.dart';
+
+abstract class GameHistoryEvent {}
+
+class SaveGameHistory extends GameHistoryEvent {
+  final GameHistoryModel gameHistory;
+
+  SaveGameHistory({required this.gameHistory});
+}
